chore(api): refactor logic into api
